4.1 apparently rounds double values different than 5.0 and 5.1.  this test only works for 5.0 and higher

+++ MySql.Data/Tests/Source/DataTypeTests.cs
@@ -960,6 +960,8 @@ namespace MySql.Data.MySqlClient.Tests
         [Test]
         public void StoringAndRetrievingDouble()
         {
+            if (version.Major < 5) return;
+
             execSQL("DROP TABLE IF EXISTS Test");
             execSQL("CREATE TABLE Test (v DOUBLE(25,20) NOT NULL)");
